[#14] refactor : 기존의 useInfiniteQuery로직을 클래스메소드로 모듈화했습니다.
